+++ demo.sh
@@ -9,33 +9,21 @@ while [ $choice -eq 3 ]; do
     read choice
     read -p "Installing required libraries. You may be prompted for a password. [enter]" y
     if [ $choice -eq 1 ] ; then
-	    if [ -f ./ve/bin/python2.7 ];
-	    then
-            sudo rm -rf ve bin
-        else
-            # Distribute and bootstrap sometimes have version conflicts. Delete.
-            sudo rm -rf ve/lib/python2.6/site-packages/distribute*
-	    fi
+        sudo rm -rf ve bin
         sudo apt-get install python-virtualenv python2.6-dev \
 	    libjpeg62-dev zlib1g-dev build-essential git-core \
-        sqlite3 libproj0 libproj-dev libgeos-3.2.2 libgdal1-dev \
+        sqlite3 libsqlite3-dev libproj0 libproj-dev libgeos-3.2.2 libgdal1-dev \
         libgdal1-1.7.0 libspatialite3 spatialite-bin libgeoip1 libgeoip-dev --no-upgrade
     	echo "Setting up sandboxed Python environment with Python 2.6"
-	    virtualenv --python=python2.6 --no-site-packages ve
+	    virtualenv --python=python2.6 --no-site-packages --setuptools ve
     else
-	    if [ -f ./ve/bin/python2.6 ];
-    	then
-            sudo rm -rf ve bin
-        else
-            # Distribute and bootstrap sometimes have version conflicts. Delete.
-            sudo rm -rf ve/lib/python2.7/site-packages/distribute*
-    	fi
+        sudo rm -rf ve bin
 	    sudo apt-get install python-virtualenv python2.7-dev \
     	libjpeg-dev zlib1g-dev build-essential git-core \
-        sqlite3 libproj0 libproj-dev libgeos-3.2.2 libgdal1-dev \
+        sqlite3 libsqlite3-dev libproj0 libproj-dev libgeos-3.2.2 libgdal1-dev \
         libgdal1-1.7.0 libspatialite3 spatialite-bin libgeoip1 libgeoip-dev --no-upgrade
 	    echo "Setting up sandboxed Python environment with Python 2.7"
-    	virtualenv --python=python2.7 --no-site-packages ve
+    	virtualenv --python=python2.7 --no-site-packages --setuptools ve
     fi
 done
 
@@ -60,7 +48,6 @@ libraries=sqlite3\
 sudo /sbin/ldconfig
 cd ..
 
-echo "Downloading distribute"
 ve/bin/python bootstrap.py -v 1.7.0
 echo "Choose the type of demo site:"
 choice=4