+++ skeleton/settings_dev_smart_site.py
@@ -3,4 +3,3 @@ from skeleton.settings import *
 
 FOUNDRY['layers'] = ('smart', 'basic')
 SITE_ID = 4
-STATIC_URL = '/static/smart/'