+++ test_settings.py
@@ -1,0 +1,15 @@
+from skeleton.settings import *
+
+
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.contrib.gis.db.backends.postgis',
+        'NAME': 'skeleton',
+        'USER': 'test',
+        'PASSWORD': '',
+        'HOST': '',
+        'PORT': '',
+    }
+}
+
+# Need this last line until django-setuptest is improved.